`ToBig`, `FromBig`, `MustFromBig`: allow conversion of nil-values (#137)
In some places, having a `nil` as a `big.Int` or `uint256.Int` might make sense (e.g. an optional and non-initialized field). In those cases, converting between the two across older/newer code might be necessary.

This changes makes it so `ToBig`, `FromBig` and `MustFromBig` do not crash, but rather just returns `nil` in case the input-to-convert is `nil`.
---------

Co-authored-by: Martin Holst Swende <martin@swende.se>