conversion: introduce alloc-free `IntoBig` method (#177)
`ToBig` does 2x32 byte allocations. Once to create a new `big.Int` and a second time to seed the contents of the `big.Int`. This change adds `IntoBig` which takes an input `big.Int`, and, if possible, it will use the given `big.Int` buffer for conversion to avoid reallocation.

If the big.Int is nil, it will be allocated (32 byte). If it's not large enough, it's data space will be allocated (32 byte).